Move traits to smartcontracts module
