+++ iroha/src/smartcontracts/isi/permissions.rs
@@ -8,7 +8,7 @@ use iroha_data_model::prelude::*;
 use iroha_error::Result;
 
 #[cfg(feature = "roles")]
-use super::expression::Evaluate;
+use super::Evaluate;
 use crate::prelude::*;
 
 /// Reason for prohibiting the execution of the particular instruction.
